+++ crates/router/src/configs/settings.rs
@@ -117,6 +117,7 @@ pub struct Settings<S: SecretState> {
     pub mandates: Mandates,
     pub zero_mandates: ZeroMandates,
     pub network_transaction_id_supported_connectors: NetworkTransactionIdSupportedConnectors,
+    pub card_only_mit_supported_connectors: CardOnlyMitSupportedConnectors,
     pub list_dispute_supported_connectors: ListDiputeSupportedConnectors,
     pub required_fields: RequiredFields,
     pub delayed_session_response: DelayedSessionConfig,
@@ -172,7 +173,6 @@ pub struct Settings<S: SecretState> {
     pub revenue_recovery: revenue_recovery::RevenueRecoverySettings,
     pub clone_connector_allowlist: Option<CloneConnectorAllowlistConfig>,
     pub merchant_id_auth: MerchantIdAuthSettings,
-    pub preprocessing_flow_config: Option<PreProcessingFlowConfig>,
     pub internal_merchant_id_profile_id_auth: InternalMerchantIdProfileIdAuthSettings,
     #[serde(default)]
     pub infra_values: Option<HashMap<String, String>>,
@@ -192,18 +192,6 @@ pub struct OnSessionConfig {
     #[serde(default, deserialize_with = "deserialize_hashmap")]
     pub unsupported_payment_methods:
         HashMap<enums::PaymentMethod, HashSet<enums::PaymentMethodType>>,
-}
-
-#[derive(Debug, Deserialize, Clone, Default)]
-pub struct PreProcessingFlowConfig {
-    #[serde(default, deserialize_with = "deserialize_hashset")]
-    pub authentication_bloated_connectors: HashSet<enums::Connector>,
-    #[serde(default, deserialize_with = "deserialize_hashset")]
-    pub order_create_bloated_connectors: HashSet<enums::Connector>,
-    #[serde(default, deserialize_with = "deserialize_hashset")]
-    pub balance_check_bloated_connectors: HashSet<enums::Connector>,
-    #[serde(default, deserialize_with = "deserialize_hashset")]
-    pub settlement_split_bloated_connectors: HashSet<enums::Connector>,
 }
 
 #[derive(Debug, Deserialize, Clone, Default)]
@@ -585,6 +573,12 @@ pub struct AuthenticationServiceEnabledConnectors {
 
 #[derive(Debug, Deserialize, Clone, Default)]
 pub struct NetworkTransactionIdSupportedConnectors {
+    #[serde(deserialize_with = "deserialize_hashset")]
+    pub connector_list: HashSet<enums::Connector>,
+}
+
+#[derive(Debug, Deserialize, Clone, Default)]
+pub struct CardOnlyMitSupportedConnectors {
     #[serde(deserialize_with = "deserialize_hashset")]
     pub connector_list: HashSet<enums::Connector>,
 }